requirements.txt
python-dotenv
pandas
numpy
pymongo
pyaml
dill
certifi
scikit-learn

##-e .